Refactor: Organize FastAPI routes using APIRouter
Moved series-related endpoints from the main server file (`src/server.py`) to a dedicated router (`src/routers/series_router.py`).

This change improves code organization and maintainability by grouping related routes. The main application now includes this router. All functionalities remain the same.
